Refactoring to clean up Spark tool invocations
Factored Spark tool calls (for conductor and ADAM) out into a
`toil_scripts.tools.spark_tools` module. In this module, I then centralized
the Spark command line creation into a module private method, so that we
weren't duplicating this code between `call_conductor` and `call_adam`.
Additionally, I added an override for this configuration code so that a user
could override all of the configuration we are passing if they are running on
an on-premises cluster where this configuration is unnecessary/invalid. I
then refactored `toil_scripts.adam_pipeline.adam_preprocessing` with these
changes.