Refactor RNASeq tests so we can call them from Toil integration tests
Added ability to run tests on a cluster, with multiple generic samples.

Also, eliminate get_work_directory() and stop passing --workDir in tests

Relying on TMPDIR is simpler and more portable. Toil already defaults to setting it's working directory based on TMPDIR.

Passing --workDir broke tests when they are run on a cluster because its value—the unique directory returned from get_work_directory()—was not valid on any workers.